requirements.txt
pytest==7.1.*
flake8
pdoc==12.*
build==0.8.*
twine
numpy
tensorflow
Pillow
setuptools
ebeer
importlib_resources
streamlit